+++ app/next-client-app/app/(protected)/scanreports/[id]/columns.tsx
@@ -7,8 +7,7 @@ import JobDialog from "@/components/jobs/JobDialog";
 import { FindGeneralStatus, DivideJobs } from "@/components/jobs/JobUtils";
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
-import CopyButton from "@/components/core/handleCopy";
-
+import CopyButton from "@/components/core/CopyButton";
 
 export const columns: ColumnDef<ScanReportTable>[] = [
   {
@@ -27,10 +26,10 @@ export const columns: ColumnDef<ScanReportTable>[] = [
               {name}
             </Button>
           </Link>
-          <CopyButton textToCopy={name} /> 
+          <CopyButton textToCopy={name} />
         </div>
       );
-    },
+    }
   },
   {
     id: "Person ID",
@@ -47,7 +46,7 @@ export const columns: ColumnDef<ScanReportTable>[] = [
       return <>{person_id?.name}</>;
     },
     enableHiding: true,
-    enableSorting: false,
+    enableSorting: false
   },
   {
     id: "Event Date",
@@ -64,7 +63,7 @@ export const columns: ColumnDef<ScanReportTable>[] = [
       return <>{date_event?.name}</>;
     },
     enableHiding: true,
-    enableSorting: false,
+    enableSorting: false
   },
   {
     id: "jobs",
@@ -88,7 +87,7 @@ export const columns: ColumnDef<ScanReportTable>[] = [
           />
         </div>
       );
-    },
+    }
   },
   {
     id: "edit",
@@ -111,6 +110,6 @@ export const columns: ColumnDef<ScanReportTable>[] = [
           generalStatus={generalStatus}
         />
       );
-    },
-  },
+    }
+  }
 ];
