+++ app/next-client-app/components/core/CopyButton.tsx
